Moved code from bouy tasks fsm to update userdata node.
